[ENG-380] Processing import in steps.
Makes it possible to recover a failure with a little manual work.

+++ Command/UpdateCorrectAddressDataCommand.php
@@ -56,7 +56,12 @@ class UpdateCorrectAddressDataCommand extends ModeratedCommand
             $output->writeln('<info>SFTP connection established, downloading data file</info>');
 
             $source = 'ssh2.sftp://'.intval($sftp).$settings[CAI::CA_REMOTE_PATH].'/'.$settings[CAI::CA_REMOTE_FILE];
-            $dest   = sys_get_temp_dir().'/ca_'.\date('Y-m-d').'.zip';
+
+            $buffer = tempnam(sys_get_temp_dir(), 'ca_'.\date('Y-m-d');
+            if (file_exists($buffer)) {
+                unlink($buffer);
+            }
+            $dest   = $buffer.'.zip';
             $rfp    = fopen($source, 'r');
             $wfp    = fopen($dest, 'w');
 
@@ -77,22 +82,35 @@ class UpdateCorrectAddressDataCommand extends ModeratedCommand
             $output->writeln('<info>Copied data archive to '.$dest.' on local filesystem.</info>');
 
             //extract the new files
-            $buffer    = '/tmp/data/eq_correct_address';
             $extractor = new ZipArchive();
             $extractor->open($dest, ZipArchive::CHECKCONS);
             $extractor->extractTo($buffer);
-            $extractor->close();
-            unlink($dest);
-            $output->writeln('<info>Archive extracted to '.$buffer.'.</info>');
+            if ($extractor->close() && is_dir($buffer)) {
+                unlink($dest);
+                $output->writeln('<info>Archive extracted to '.$buffer.'.</info>');
 
-            //remove the old files
-            $this->cleanDir($settings[CAI::CA_CORRECTA_DATA]);
-            $output->writeln('<info>'.$settings[CAI::CA_CORRECTA_DATA].' cleaned.</info>');
+                if (is_dir($settings[CAI::CA_CORRECTA_DATA].'_bak')) {
+                    $this->cleanDir($settings[CAI::CA_CORRECTA_DATA].'_bak');
+                    $output->writeln('<info>'.$settings[CAI::CA_CORRECTA_DATA].'_bak cleaned.</info>');
+                }
+                rename($settings[CAI::CA_CORRECTA_DATA], $settings[CAI::CA_CORRECTA_DATA].'_bak');
 
-            rename($buffer, $settings[CAI::CA_CORRECTA_DATA]);
-            $output->writeln('<info>Expirian data update complete.</info>');
+                if (rename($buffer, $settings[CAI::CA_CORRECTA_DATA])) {
+                    $output->writeln('<info>Expirian data update complete.</info>');
 
-            return 0;
+                    return 0;
+                } else {
+                    $output->writeln(
+                        '<error>Failed to move '.$buffer.' to '.$settings[CAI::CA_CORRECTA_DATA].'</error>'
+                    );
+
+                    return 1;
+                }
+            } else {
+                $output->writeln('<error>Failed to unzip '.$dest.' to '.$buffer.'</error>');
+
+                return 1;
+            }
         } catch (\Exception $e) {
             $output->writeln('<error>Failed to update data: '.$e->getMessage().'</error>');
             $output->write($e->getTraceAsString());
